+++ src/routes/timer.svelte
@@ -1,6 +1,5 @@
 <script>
   import Timer from "../components/Timer.svelte";
-  import Timer2 from "../components/Timer2.svelte";
 </script>
 
 <svelte:head>
@@ -9,5 +8,4 @@
 
 <h1>Exercise Timer</h1>
 
-<!-- <Timer /> -->
-<Timer2 />
+<Timer />
